+++ guillotina/content.py
@@ -647,8 +647,8 @@ async def get_all_behaviors(content, create=False, load=True) -> list:
 
 async def duplicate(
     context: IResource,
-    destination: Union[IResource, str] = None,
-    new_id: str = None,
+    destination: Optional[Union[IResource, str]] = None,
+    new_id: Optional[str] = None,
     check_permission: bool = True,
 ) -> IResource:
     if destination is not None:
@@ -700,6 +700,7 @@ async def duplicate(
         id=new_id,
         creators=context.creators,
         contributors=context.contributors,
+        check_security=check_permission,
     )
 
     for key in context.__dict__.keys():
@@ -732,8 +733,8 @@ async def duplicate(
 
 async def move(
     context: IResource,
-    destination: Union[IResource, str] = None,
-    new_id: str = None,
+    destination: Optional[Union[IResource, str]] = None,
+    new_id: Optional[str] = None,
     check_permission: bool = True,
 ) -> None:
     if destination is None: